+++ src/main/cpp/auton/primitives/IPrimitive.h
@@ -21,7 +21,7 @@ class PrimitiveParams;
 class IPrimitive
 {
     public:
-         const float LOOP_LENGTH = 0.020; //Program loop duration in seconds
+         const double LOOP_LENGTH = 0.020; //Program loop duration in seconds
 
         IPrimitive() = default;
         virtual ~IPrimitive() = default;
